installer: fetch one tag after doing a shallow clone
When the installer downloads a release branch, creation of local IDF
repository is done in two stages:

1. Download the zip file of one of the releases on the same branch,
   extract it in a temporary directory.
2. Perform git clone, using the extracted directory as reference.

In the 2nd step, we may perform "shallow" clone if such option is
chosen by the user. Shallow repository doesn't contain any tags, so
"git describe" will later not work correctly in the repository.
In particular, IDF build system will not determine the version
correctly, which can be seen in the output of the 2nd stage
bootloader.

This is a partial fix to the problem. In order to make "git describe"
work, we fetch one of the tags from the same release branch.

Note that this fix still doesn't solve the issue in two cases:
- If no release tag exists for this branch. For example, release/v5.1
  has been already created and added to idf_versions.txt, but no
  tagged release is available yet.
- When cloning the master branch.